REWORD PHP

+++ php/delete.php
@@ -1,8 +1,17 @@
 <?php
-include_once 'db.php';
+$servername = "localhost";
+$username = "root";
+$password = "123";
+$dbname = "smoer";
+
+$conn = mysqli_connect($servername, $username, $password, $dbname);
+ if (!$conn) {
+     die("Connection failed: " . mysqli_connect_error());
+ }
+
 $id = $_GET['id'];
 
-$sql = "DELETE FROM news WHERE id= '$id'   ";
+$sql = "DELETE FROM news WHERE id= '$id' ";
 
 if (mysqli_query($conn, $sql)) {
     header('location: admin.php');
